// src/core/interfaces/user.interface.ts
/**
 * Copyright (c) 2025 Webapp Software Solutions. All rights reserved.
 * This file belongs to Webapp Software Solutions and is proprietary and confidential.
 * Unauthorized copying of this file, via any medium is strictly prohibited.
 * Website: https://webappssoft.com
 */

import type { UserRoles } from '../../shared/enums/accouts.enum';

export interface IUser {
  id: string;
  settingId: string;
  role: UserRoles;
}